Rename AlmightyContentModule to ImageManagementModule
