+++ Kitodo-ImageManagementModule/pom.xml
@@ -21,8 +21,8 @@
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <name>Kitodo − Almighty Content Module</name>
-    <artifactId>kitodo-almightycontentmodule</artifactId>
+    <name>Kitodo − Image Management Module</name>
+    <artifactId>kitodo-ImageManagementModule</artifactId>
 
     <dependencies>
         <dependency>
