+++ Kitodo-ImageManagementModule/src/main/java/org/kitodo/imagemanagementmodule/ImageConverter.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package org.kitodo.almightycontentmodule;
+package org.kitodo.imagemanagementmodule;
 
 import java.io.File;
 import java.io.IOException;
