+++ pom.xml
@@ -111,9 +111,9 @@
         <module>Kitodo-DataManagement</module>
         <module>Kitodo-DataAccess</module>
         <module>Kitodo-API</module>
-        <module>Kitodo-AlmightyContentModule</module>
         <module>Kitodo-DataEditor</module>
         <module>Kitodo-DataFormat</module>
+        <module>Kitodo-ImageManagementModule</module>
         <module>Kitodo-PersistentIdentifier</module>
         <module>Kitodo-Docket</module>
         <module>Kitodo-Command</module>
